backend/blockcontainer/requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
beautifulsoup4==4.6.1
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
Django==2.0.7
django-celery==3.2.2
django-celery-results==1.0.1
django-node==4.0.2
django-rest-framework==0.1.0
django-rest-knox==3.1.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
feedparser==5.2.1
idna==2.7
kombu==4.2.1
optional-django==0.3.0
packaging==17.1
psycopg2==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
pyparsing==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
semantic-version==2.6.0
six==1.11.0
updatable==0.3.1
urllib3==1.23
vine==1.1.4
